adding julia version to windows

+++ .artenolis/runtests.sh
@@ -5,7 +5,6 @@ echo "JULIA_VER = $JULIA_VER"
 # launch the test script
 if [ "$ARCH" == "Linux" ]; then
     if [ "$JULIA_VER" == "v0.6.4" ]; then
-
         # remove th julia directory to clean the installation directory
         rm -rf ~/.julia/v0.6/COBRA
 
@@ -18,17 +17,19 @@ if [ "$ARCH" == "Linux" ]; then
     fi
 
 elif [ "$ARCH" == "windows" ]; then
-    # remove th julia directory to clean the installation directory
-    rm -rf ~/.julia/v0.6/COBRA
+    if [ "$JULIA_VER" == "v0.6.4" ]; then
+        # remove th julia directory to clean the installation directory
+        rm -rf ~/.julia/v0.6/COBRA
 
-    unset Path
-    nohup "D:\\JULIA\\$JULIA_VER\\\bin\\julia.exe" --color=yes -e 'Pkg.clone(pwd()); cd(Pkg.dir("COBRA")); Pkg.test(pwd());' > output.log & PID=$!
+        unset Path
+        nohup "D:\\JULIA\\$JULIA_VER\\\bin\\julia.exe" --color=yes -e 'Pkg.clone(pwd()); cd(Pkg.dir("COBRA")); Pkg.test(pwd());' > output.log & PID=$!
 
-    # follow the log file
-    tail -n0 -F --pid=$! output.log 2>/dev/null
+        # follow the log file
+        tail -n0 -F --pid=$! output.log 2>/dev/null
 
-    # wait until the background process is done
-    wait $PID
+        # wait until the background process is done
+        wait $PID
+    fi
 fi
 
 CODE=$?
